Fix HttpReponse mapping
