+++ tests/NClient.Testing/NClient.Testing.Common/NClient.Testing.Common.csproj
@@ -32,6 +32,7 @@
     </ItemGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\..\..\src\NClient.Providers\NClient.Providers.Mapping.HttpResponses\NClient.Providers.Mapping.HttpResponses.csproj" />
         <ProjectReference Include="..\..\..\src\NClient\NClient.Annotations\NClient.Annotations.csproj" />
         <ProjectReference Include="..\..\..\src\NClient\NClient.Core\NClient.Core.csproj" />
         <ProjectReference Include="..\..\..\src\NClient\NClient.Standalone\NClient.Standalone.csproj" />
